Molecule -> MoleculeEditor, MoleculeRepr -> Molecule
