+++ crates/molecule/src/lib.rs
@@ -1,7 +1,9 @@
-pub use crate::molecule::{AtomIndex, BondIndex, BondOrder, Molecule, MoleculeGraph};
+pub use crate::molecule::{AtomIndex, BondIndex, BondOrder, MoleculeGraph};
+pub use crate::molecule_editor::MoleculeEditor;
 
 mod dynamics;
 pub mod edit;
 mod molecule;
+mod molecule_editor;
 mod pdb;
 mod vsepr;
