+++ crates/molecule/src/molecule.rs
@@ -8,9 +8,14 @@ use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 use ultraviolet::Vec3;
 
-use crate::edit::{Edit, EditContext, EditError, EditList, ReferenceType};
+use crate::edit::{EditContext, EditError, ReferenceType};
 
 use common::{ids::AtomSpecifier, BoundingBox};
+
+lazy_static! {
+    pub static ref PERIODIC_TABLE: periodic_table::PeriodicTable =
+        periodic_table::PeriodicTable::new();
+}
 
 pub type MoleculeGraph = stable_graph::StableUnGraph<AtomNode, BondOrder>;
 // A map that gives each atom in a molecule a coordinate. Used to cache structure energy minimization
@@ -61,7 +66,7 @@ impl AtomNode {
 
 /// The concrete representation of the molecule at some time in the feature history.
 #[derive(Default)]
-pub struct MoleculeRepr {
+pub struct Molecule {
     // TODO: This atom map is a simple but extremely inefficient implementation. This data
     // is highly structued and repetitive: compression, flattening, and a tree could do
     // a lot to optimize this.
@@ -73,8 +78,8 @@ pub struct MoleculeRepr {
     positions: AtomPositions,
 }
 
-impl MoleculeRepr {
-    fn atom_reprs(&self) -> Vec<AtomRepr> {
+impl Molecule {
+    pub fn atom_reprs(&self) -> Vec<AtomRepr> {
         self.graph
             .node_weights()
             .map(|node| AtomRepr {
@@ -86,7 +91,7 @@ impl MoleculeRepr {
             .collect()
     }
 
-    fn clear(&mut self) {
+    pub fn clear(&mut self) {
         self.atom_map.clear();
         self.graph.clear();
         self.bounding_box = Default::default();
@@ -133,14 +138,63 @@ impl MoleculeRepr {
             positions: self.positions.clone(),
         }
     }
+
+    pub fn bounding_box(&self) -> &BoundingBox {
+        &self.bounding_box
+    }
+
+    // TODO: Optimize heavily (use octree, compute entry point of ray analytically)
+    pub fn get_ray_hit(&self, origin: Vec3, direction: Vec3) -> Option<AtomSpecifier> {
+        // Using `direction` as a velocity vector, determine when the ray will
+        // collide with the bounding box. Note the ? - this fn returns early if there
+        // isn't a collision.
+        let (tmin, tmax) = self.bounding_box.ray_hit_times(origin, direction)?;
+
+        // If the box is fully behind the raycast direction, we will never get a hit.
+        if tmax <= 0.0 {
+            return None;
+        }
+
+        // Knowing that the ray will enter the box, we can now march along it by a fixed step
+        // size. At each step, we check for a collision with an atom, and return that atom's index
+        // if a collision occurs.
+
+        // We know that the box is first hit at `origin + tmin * direction`. However,
+        // tmin can be negative, and we only want to march forwards. So,
+        // we constrain tmin to be nonnegative.
+        let mut current_pos = origin + f32::max(0.0, tmin) * direction;
+
+        // This is an empirically reasonable value. It is still possible to miss an atom if
+        // the user clicks on the very edge of it, but this is rare.
+        let step_size = PERIODIC_TABLE.element_reprs[Element::Hydrogen as usize].radius / 10.0;
+        let step = direction * step_size;
+        let t_span = tmax - f32::max(0.0, tmin);
+        // the direction vector is normalized, so 1 unit of time = 1 unit of space
+        let num_steps = (t_span / step_size) as usize;
+
+        for _ in 0..num_steps {
+            for atom in self.graph.node_weights() {
+                let atom_radius_sq = PERIODIC_TABLE.element_reprs[atom.element as usize]
+                    .radius
+                    .powi(2);
+
+                let atom_pos = *self
+                    .positions
+                    .get(&atom.spec)
+                    .expect("Every atom in the graph should have an associated position");
+                if (current_pos - atom_pos).mag_sq() < atom_radius_sq {
+                    return Some(atom.spec.clone());
+                }
+            }
+
+            current_pos += step;
+        }
+
+        None
+    }
 }
 
-lazy_static! {
-    pub static ref PERIODIC_TABLE: periodic_table::PeriodicTable =
-        periodic_table::PeriodicTable::new();
-}
-
-impl EditContext for MoleculeRepr {
+impl EditContext for Molecule {
     fn add_bonded_atom(
         &mut self,
         element: Element,
@@ -206,261 +260,5 @@ impl EditContext for MoleculeRepr {
 
     fn pos(&self, spec: &AtomSpecifier) -> Option<&Vec3> {
         self.positions.get(spec)
-    }
-}
-
-/// Demonstration of how to use the feature system
-/// let mut molecule = Molecule::from_feature(
-///     &gpu_resources,
-///     RootAtom {
-///         element: Element::Iodine,
-///     },
-/// );
-///
-/// molecule.push_feature(AtomFeature {
-///     target: scene::ids::AtomSpecifier::new(0),
-///     element: Element::Sulfur,
-/// });
-/// molecule.apply_all_features();
-///
-/// molecule.push_feature(AtomFeature {
-///     target: scene::ids::AtomSpecifier::new(1),
-///     element: Element::Carbon,
-/// });
-/// molecule.apply_all_features();
-///
-/// molecule.set_history_step(2);
-/// molecule.reupload_atoms(&gpu_resources);
-pub struct Molecule {
-    pub repr: MoleculeRepr,
-    #[allow(unused)]
-    rotation: ultraviolet::Rotor3,
-    #[allow(unused)]
-    offset: ultraviolet::Vec3,
-    edits: EditList,
-    // The index one greater than the most recently applied feature's location in the feature list.
-    // This is unrelated to feature IDs: it is effectively just a counter of how many features are
-    // applied. (i.e. our current location in the edit history timeline)
-    history_step: usize,
-    // when `history_step` is set to `i`, if `checkpoints` contains the key `i`, then
-    // `checkpoints.get(i)` contains the graph and geometry which should be used to render
-    // the molecule. This allows feature application and relaxation to be cached until they
-    // need to be recomputed. This saves a lot of time, as relaxation is a very expensive operation that does not
-    // commute with feature application.
-    checkpoints: HashMap<usize, MoleculeCheckpoint>,
-    // the history step we cannot equal or exceed without first recomputing. For example, if repr
-    // is up to date with the feature list, and then a past feature is changed, dirty_step would change
-    // from `features.len()` to the index of the changed feature. This is used to determine if recomputation
-    // is needed when moving forwards in the timeline, or if a future checkpoint can be used.
-    dirty_step: usize,
-}
-
-impl Molecule {
-    pub fn from_feature(edit: Edit) -> Self {
-        let mut repr = MoleculeRepr::default();
-        edit.apply(&0, &mut repr)
-            .expect("Primitive features should never return a feature error!");
-        repr.relax();
-
-        let mut features = EditList::default();
-        features.push_back(edit);
-
-        Self {
-            repr,
-            rotation: ultraviolet::Rotor3::default(),
-            offset: ultraviolet::Vec3::default(),
-            edits: features,
-            history_step: 1, // This starts at 1 because we applied the primitive feature
-            checkpoints: Default::default(),
-            dirty_step: 1, // Although no checkpoints exist, repr is not dirty, so we advance this to its max
-        }
-    }
-
-    pub fn edits(&self) -> &EditList {
-        &self.edits
-    }
-
-    pub fn insert_edit(&mut self, edit: Edit) {
-        self.edits.insert(edit, self.history_step);
-    }
-
-    // Advances the model to a given history step by applying features in the timeline.
-    // This will not in general recompute the history, so if a past feature is changed,
-    // you must recompute from there.
-    pub fn set_history_step(&mut self, history_step: usize) {
-        // TODO: Bubble error to user
-        assert!(
-            history_step <= self.edits.len(),
-            "history step exceeds edit list size"
-        );
-
-        // Find the best checkpoint to start reconstructing from:
-        let best_checkpoint = self
-            .checkpoints
-            .keys()
-            .filter(|candidate| **candidate <= history_step)
-            .max();
-
-        match best_checkpoint {
-            None => {
-                // If there wasn't a usable checkpoint, we can either keep computing forwards or
-                // restart. We only have to restart from scratch if we're moving backwards, otherwise
-                // we can just move forwards.
-
-                if self.history_step > history_step {
-                    self.history_step = 0;
-                    self.repr.clear();
-                }
-            }
-            Some(best_checkpoint) => {
-                // If there was, we can go there and resume from that point
-                self.repr
-                    .set_checkpoint(self.checkpoints.get(best_checkpoint).unwrap().clone());
-                self.history_step = *best_checkpoint;
-            }
-        }
-
-        for edit_id in &self.edits.order()[self.history_step..history_step] {
-            println!("Applying edit {}", edit_id);
-            let edit = self
-                .edits
-                .get(edit_id)
-                .expect("Feature IDs referenced by the FeatureList order should exist!");
-
-            if edit.apply(edit_id, &mut self.repr).is_err() {
-                // TODO: Bubble error to the user
-                println!("Failed to apply the edit with id {}", edit_id);
-                dbg!(&edit);
-            }
-
-            self.repr.relax();
-        }
-
-        self.dirty_step = history_step;
-        self.history_step = history_step;
-    }
-
-    // equivalent to `set_history_step(features.len()): applies every feature that is in the
-    // feature timeline.
-    pub fn apply_all_edits(&mut self) {
-        self.set_history_step(self.edits.len())
-    }
-
-    // TODO: Optimize heavily (use octree, compute entry point of ray analytically)
-    pub fn get_ray_hit(&self, origin: Vec3, direction: Vec3) -> Option<AtomSpecifier> {
-        // Using `direction` as a velocity vector, determine when the ray will
-        // collide with the bounding box. Note the ? - this fn returns early if there
-        // isn't a collision.
-        let (tmin, tmax) = self.repr.bounding_box.ray_hit_times(origin, direction)?;
-
-        // If the box is fully behind the raycast direction, we will never get a hit.
-        if tmax <= 0.0 {
-            return None;
-        }
-
-        // Knowing that the ray will enter the box, we can now march along it by a fixed step
-        // size. At each step, we check for a collision with an atom, and return that atom's index
-        // if a collision occurs.
-
-        // We know that the box is first hit at `origin + tmin * direction`. However,
-        // tmin can be negative, and we only want to march forwards. So,
-        // we constrain tmin to be nonnegative.
-        let mut current_pos = origin + f32::max(0.0, tmin) * direction;
-
-        // This is an empirically reasonable value. It is still possible to miss an atom if
-        // the user clicks on the very edge of it, but this is rare.
-        let step_size = PERIODIC_TABLE.element_reprs[Element::Hydrogen as usize].radius / 10.0;
-        let step = direction * step_size;
-        let t_span = tmax - f32::max(0.0, tmin);
-        // the direction vector is normalized, so 1 unit of time = 1 unit of space
-        let num_steps = (t_span / step_size) as usize;
-
-        let graph = &self.repr.graph;
-        for _ in 0..num_steps {
-            for atom in graph.node_weights() {
-                let atom_radius_sq = PERIODIC_TABLE.element_reprs[atom.element as usize]
-                    .radius
-                    .powi(2);
-
-                let atom_pos = *self
-                    .repr
-                    .positions
-                    .get(&atom.spec)
-                    .expect("Every atom in the graph should have an associated position");
-                if (current_pos - atom_pos).mag_sq() < atom_radius_sq {
-                    return Some(atom.spec.clone());
-                }
-            }
-
-            current_pos += step;
-        }
-
-        None
-    }
-}
-
-// This is a stripped down representation of the molecule that removes several
-// fields (some are redundant, like repr.atom_map, and some are not serializable,
-// like repr.gpu_atoms).
-#[derive(Serialize, Deserialize)]
-struct ProxyMolecule {
-    rotation: ultraviolet::Rotor3,
-    offset: ultraviolet::Vec3,
-    edits: EditList,
-    history_step: usize,
-    checkpoints: HashMap<usize, MoleculeCheckpoint>,
-    dirty_step: usize,
-}
-
-impl Serialize for Molecule {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        // Custom serialization is used to ensure that the current molecule state is
-        // saved as a checkpoint, even if it normally would not be (i.e. if it's already
-        // very close to an existing checkpoint). This allows faster loading when the file
-        // is reopened.
-
-        let mut checkpoints = self.checkpoints.clone();
-        checkpoints.insert(self.history_step, self.repr.make_checkpoint());
-
-        let data = ProxyMolecule {
-            rotation: self.rotation,
-            offset: self.offset,
-            edits: self.edits.clone(),
-            history_step: self.history_step,
-            checkpoints,
-            dirty_step: self.dirty_step,
-        };
-
-        data.serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for Molecule {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        // TODO: integrity check of the deserialized struct
-
-        let data = ProxyMolecule::deserialize(deserializer)?;
-
-        let mut molecule = Molecule {
-            repr: MoleculeRepr::default(),
-            rotation: data.rotation,
-            offset: data.offset,
-            edits: data.edits,
-            history_step: data.history_step, // This starts at 0 because we haven't applied the features, we've just loaded them
-
-            checkpoints: data.checkpoints,
-            dirty_step: data.dirty_step,
-        };
-
-        // this advances the history step to the correct location
-        molecule.set_history_step(data.history_step);
-
-        Ok(molecule)
     }
 }
